setup code structure, init peano

+++ src/TreeNodes.jl
@@ -49,7 +49,7 @@ mutable struct TopNode{I<:Integer}
     StartKey::Int128
     Count::Int128
 end
-TopNode(;bits=21) = TopNode(-1, 1, 0, 0, Int128(1)<<Int128(3*bits), Int128(0), Int128(0))
+TopNode(;bits=21) = TopNode(-1, 1, 0, 0, Int128(1)<<(3*bits), Int128(0), Int128(0))
 
 mutable struct PhysicalOctreeNode2D{I<:Integer} <: AbstractOctreeNode2D{I}
     ID::I
